fix: change id integer type to long

+++ src/main/java/br/edu/fema/modelo/atividadesfixacao/domain/alimento/entity/TipoAlimento.java
@@ -14,7 +14,7 @@ public class TipoAlimento {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Setter(AccessLevel.NONE)
-    private Integer id; // ta long no Repository
+    private long id; // ta long no Repository
 
     @Column(length = 20, nullable = false)
     private String descricao;
